[MIG] pos_product_multi_barcode: Migration to 17.0
[DON'T MERGE] test-requirements.txt
